Basic support for bluepill board with at86rf233 radio
The bluepill board[1] is a tiny arduino-like developer board
that is equipped with a powerful STM32F103 MCU. The chip
is officially an STM32F103C8T6 model that in fact appears
to have 128 KiB flash (cf. [2]). Thus it is recommended to
use the patched stlink[3] to flash the firmware.

[1] http://wiki.stm32duino.com/index.php?title=Blue_Pill
[2] https://github.com/texane/stlink/issues/305
[3] https://github.com/cabo/stlink